backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-multipart==0.0.6
google-api-python-client==2.108.0
google-generativeai==0.8.3
python-dotenv==1.0.0
textblob==0.17.1
nltk==3.8.1
scikit-learn==1.3.2
langdetect==1.0.9
numpy==1.24.3
matplotlib==3.7.2
Pillow==10.0.1
